fix(validator): primitive iml expression bypass (#18)
In case a field contains a primitive IML Expression, the primitive type
validators for number, boolean and so on are currently mistakenly
returning validation error, even when a valid-typed primitive value
could be mapped into them. This is patch for such behavior.

_There's a catch when someone would do:_

```
const context = {
    first: 'fal',
    second: 'se'
}
```

and then `{{context.first}}{{context.second}}` to get `false` as string
which would be casted later, similarly with numbers, but that sounds too
obscure to be significant at the moment.

If we ever encounter a problem with the `isPrimitiveIMLExpression`, we
can always make it more relaxed to the `containsIMLExpression` instead,
but we want to be stricter when we can now.

---------

Co-authored-by: Patrik Simek <patrik@patriksimek.cz>